+++ src/util/path.rs
@@ -12,14 +12,15 @@ pub fn current_dir() -> Result<PathBuf> {
     env::current_dir().context("could not get current directory")
 }
 
-// Extracts the filename from a path. Returns an empty string if not found.
+/// Lexically extracts the filename from a path. Returns an empty string if not
+/// found.
 pub fn filename_str<S: AsRef<str>>(path: &S) -> &str {
     let mut path = path.as_ref();
     if cfg!(unix) {
-        if path.ends_with('/') {
-            path = &path[..path.len() - 1];
-        }
-        match path.rfind('/') {
+        const SEPARATOR: char = '/';
+        path = path.trim_end_matches(SEPARATOR);
+        // NOTE: can be written using str::rsplit_once once it is stabilized.
+        match path.rfind(SEPARATOR) {
             Some(idx) => &path[idx + 1..],
             None => path,
         }
@@ -32,15 +33,36 @@ pub fn filename_str<S: AsRef<str>>(path: &S) -> &str {
     }
 }
 
-// Resolves all path components lexically (without accessing the filesystem).
-//
-// This is similar to `realpath -ms`, except that it returns the current
-// directory when given an empty path, whereas realpath would return
-// "ENOENT: no such file or directory".
-//
-// This function is only available on UNIX systems. Since Windows filesystems
-// are typically case-insensitive but case-preserving, normalization would not
-// work there.
+/// Determines if zoxide can support the given path. It is assumed that the
+/// path is a completely resolved absolute path.
+///
+/// - On UNIX, all paths are supported.
+/// - On Windows, all Win32 drive absolute paths (`C:\foo\bar`) are supported.
+///   <https://googleprojectzero.blogspot.com/2016/02/the-definitive-guide-on-win32-to-nt.html>
+///   <https://docs.microsoft.com/en-us/windows/win32/fileio/naming-a-file>
+pub fn is_supported<P: AsRef<Path>>(path: P) -> bool {
+    let path = path.as_ref();
+    if cfg!(windows) {
+        use std::path::{Component, Prefix};
+        let mut components = path.components();
+        match components.next() {
+            Some(Component::Prefix(prefix)) if matches!(prefix.kind(), Prefix::Disk(_)) => {
+                components.next() == Some(Component::RootDir)
+            }
+            _ => return false,
+        }
+    } else {
+        true
+    }
+}
+
+/// Resolves all path components lexically (without accessing the filesystem).
+///
+/// This is similar to `realpath -ms`, except that it returns the current
+/// directory when given an empty path, whereas `realpath` would return
+/// `ENOENT: no such file or directory`.
+///
+/// This function is only available on UNIX.
 #[cfg(unix)]
 pub fn normalize<P: AsRef<Path>>(path: P) -> Result<PathBuf> {
     use std::ffi::OsString;
@@ -90,14 +112,30 @@ pub fn to_str<P: AsRef<Path>>(path: &P) -> Result<&str> {
 
 #[cfg(test)]
 mod tests {
-    use std::path::Path;
+    #[cfg(unix)]
+    #[test]
+    fn test_filename_str() {
+        const TEST_CASES: &[(&str, &str)] = &[
+            ("", ""),
+            ("/", ""),
+            ("//", ""),
+            ("///", ""),
+            ("///foo/.//bar//", "bar"),
+            ("///foo/.//bar//.//..//.//baz", "baz"),
+        ];
+        for &(input, expected) in TEST_CASES {
+            assert_eq!(super::filename_str(&input), expected)
+        }
+    }
 
     #[cfg(unix)]
     #[test]
     fn test_normalize() {
+        use std::os::unix::ffi::OsStrExt;
+
         const TEST_CASES: &[(&str, &str)] = &[
             ("/", "/"),
-            ("//", "//"),
+            ("//", "/"),
             ("///", "/"),
             ("///foo/.//bar//", "/foo/bar"),
             ("///foo/.//bar//.//..//.//baz", "/foo/baz"),
@@ -106,8 +144,11 @@ mod tests {
             ("/a/b/c/../../../x/y/z", "/x/y/z"),
             ("///..//./foo/.//bar", "/foo/bar"),
         ];
-        for (input, expected) in TEST_CASES {
-            assert_eq!(super::normalize(input).unwrap(), Path::new(expected))
+        for &(input, expected) in TEST_CASES {
+            assert_eq!(
+                super::normalize(input).unwrap().as_os_str().as_bytes(),
+                expected.as_bytes(),
+            )
         }
 
         // Empty path should not return an error.